Assignment complete
My last commit actually implemented the entire app correctly, but a lot of the tests were failing, hence this commit. Most of this commit's code references the solution code, since I was unsure how specific the spec was. It turns out that one of the spec's dependencies (enzyme) was too outdated and not configured properly; I updated enzyme and got all but three of the tests to pass. The last three tests required some configuration/set up, and considering that I still have work to do on my FEC and finished all the Redux-relevant material, I've decided to end it here.